+++ plugins/prometheus/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.3.4
+* Prometheus bug fix
+## 1.3.3
+* storage metric bug fix
+## 1.3.2
+* fixing network transmit and receive
+
 ## 1.3.1
 * update dependencies